requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiogram==3.4.1
python-socketio==5.11.2
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
psutil==5.9.5
uptime==3.0.1
cachelib==0.9.0
Werkzeug==3.0.3